requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
lxml==3.7.3
Markdown==2.6.8
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0

